updated Readme file & added type definition comments
